Improved event output formatting
`ical2org/event-format' now handles less fields: SUMMARY, DESCRIPTION and LINES.

The third field `LINES' is special. It specifies formatting of several
icalendar fields, one per line. Formatting of `LINES' is specified in the new
user option `ical2org/event-line-format'.

This user option holds a list of pointed-pair.

Each pair binds a field of the `ical2org/event' structure to a formatting
string. When the requested field is not set in the event entry, associated
output formatting string is not emitted.